refactor move resources
