+++ processor/src/db/models/coin_models/coin_infos.rs
@@ -35,7 +35,7 @@ impl CoinInfo {
         txn_timestamp: chrono::NaiveDateTime,
         wsc_index: i64,
     ) -> anyhow::Result<Option<Self>> {
-        match &CoinResource::from_write_resource(write_resource, txn_version)? {
+        match &CoinResource::from_write_resource(write_resource, txn_version, txn_timestamp)? {
             Some(CoinResource::CoinInfoResource(inner)) => {
                 let coin_info_type = &CoinInfoType::from_move_type(
                     &write_resource.r#type.as_ref().unwrap().generic_type_params[0],
