+++ processor/src/db/models/fungible_asset_models/v2_fungible_metadata.rs
@@ -111,7 +111,7 @@ impl FungibleAssetMetadataModel {
         txn_version: i64,
         txn_timestamp: chrono::NaiveDateTime,
     ) -> anyhow::Result<Option<Self>> {
-        match &CoinResource::from_write_resource(write_resource, txn_version)? {
+        match &CoinResource::from_write_resource(write_resource, txn_version, txn_timestamp)? {
             Some(CoinResource::CoinInfoResource(inner)) => {
                 let coin_info_type = &CoinInfoType::from_move_type(
                     &write_resource.r#type.as_ref().unwrap().generic_type_params[0],
