+++ processor/src/db/models/old_default_models/mod.rs
@@ -1,5 +1,4 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod postgres_move_resources;
 pub mod postgres_move_tables;
